requirements.txt
python-telegram-bot
python-dotenv
mysql-connector-python